feat: Add code2words( code_dict, codeset) to lm_helpers.py
To find words that sound like each other, we first need to create
equivalence classes of words that sound like each other (have the same
Metaphone code).  Later, we will use this to change find "sound
neighborhoods" - words whose codes are within a certain edit distance
of the code of a given word.